requirements.txt
dj-database-url
django-extensions
mezzanine
mysqlclient
prettyconf
